Dumb Liquid Distribution In
Doesn't have a way to pull liquids out of tanks - relies on liquid being
pushed into it.  Doesn't check to see if neighboring tanks contain the
same liquid - thus, ratios will be thrown off by liquids not of the type
in the pipe.  Will attempt to fill tanks containing liquids not in the
pipe.