Whew.  MicroFurnace is fully(minus one core functionality) working
I can't seem to get it to save its currently smelting item stack.  Hm.
Maybe I can have it store the ID of its stack, and then recreate it once
it loads?  Anyways, Forge doesn't seem to be able to find GUIs the same
way, to get the full glory of this push you'll have to manually add
MicroFurnace.png to the mods folder inside the minecraft.jar  .  This
one was a doozy, but I'm glad I know how to do it.