Added Basic Textures to Furnace
Also implemented lit version of furnace.
TODO:  Find a lighting function that actually takes the light of the
block into account?!?
Find a fix/workaround for wire connections not registering with renderer
until Updated.  (Buildcraft bypassed the block's array and had the
renderer figure it out as it goes, might have to do the same, seems
dreadfully inefficient though)
Fix smeltingItem not being stored in NBT.