FemtocraftResearch should now save to a file.
Research is saved Server-side, separate from PlayerData.  This is to
allow for the possibility of player-linked research advantages from
TileEntities or items, such as the possibility of Item Craftability
being linked to TileEntities instead of crafting player's research
level, or to allow TileEntities placed by a higher research level player
to have better attributes, even while the player is offline.