Due to somewhat overzealous reflection in attempting to be elegant, inelegance now ensues.
Due to Forge's weird and draconian enforcement of sidedness, attempting to load a class on the wrong side, instead of simply returning no class, throws an uncatchable exception.  Thus, I have to filter out via class and package names everything that might want to be loaded on a client, and enforce it to not be, via the configurable class finder.  Not too bad, until it starts happening when I reference certain client things from tile entities.  Like sound.