Fix issue #34.
Error was that Minecraft containers, in their infinite wisdom, ignore the passed in slot index and instead refer to an internal reference that is privately incremented when a slot is added to the container.  Thus, because the schematic/output slots were added before the reconstructor slots, those display slots were incremented by two.  So when attempting to merge the itemstack into slot 10, it was referring to the last slot of the reconstructing displaySlots, which don't accept items, thus voiding them.