Fix ssp owner research not saving.
Issue was shared ResearchManager between Client/Server on ssp.  Old sync implementation replaced PlayerResearch, which overwrote the Server-generated one which had a File reference.  The client does not, preventing saving on the client, but this carried over to server.

New implementation loads into existing PlayerResearch, if found.  If not, reverts to old behavior and simply assigns.